Refactor: Improve code maintainability
Refactor code to improve maintainability.
